+++ py_gnome/gnome/basic_types.py
@@ -4,7 +4,7 @@ from gnome import c_gnome
 
 
 """ Basic types. Should each have a corresponding, equivalent type in the Cython."""
-world_point_type = numpy.float32
+world_point_type = numpy.float64
 world_point = numpy.dtype([('p_long', world_point_type),
                            ('p_lat', world_point_type)],
                            align=True)
